task12-refactoring

+++ feature/screen/note/src/commonMain/kotlin/ru/glyph/screen/note/ui/NoteScreenViewModel.kt
@@ -2,7 +2,9 @@ package ru.glyph.screen.note.ui
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import kotlinx.coroutines.DelicateCoroutinesApi
 import kotlinx.coroutines.FlowPreview
+import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.debounce
@@ -10,9 +12,16 @@ import kotlinx.coroutines.flow.distinctUntilChanged
 import kotlinx.coroutines.flow.filterIsInstance
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.launch
+import org.jetbrains.compose.resources.stringResource
 import ru.glyph.database.api.NotesRepository
 import ru.glyph.navigation.api.Navigator
+import ru.glyph.navigation.api.model.BottomSheet
 import ru.glyph.screen.note.ui.state.NoteUiState
+import ru.glyph.string.resources.Res
+import ru.glyph.string.resources.note_delete_confirmation
+import ru.glyph.string.resources.profile_sign_out_confirmation
+import ru.glyph.utils.flow.collectIn
+import kotlin.time.Duration.Companion.seconds
 
 @OptIn(FlowPreview::class)
 internal class NoteScreenViewModel(
@@ -34,16 +43,14 @@ internal class NoteScreenViewModel(
             )
         }
 
-        viewModelScope.launch {
-            _uiState
-                .filterIsInstance<NoteUiState.Editing>()
-                .map { it.title to it.content }
-                .distinctUntilChanged()
-                .debounce(700)
-                .collect { (title, content) ->
-                    notesRepository.update(noteId, title, content)
-                }
-        }
+        uiState
+            .filterIsInstance<NoteUiState.Editing>()
+            .map { it.title to it.content }
+            .distinctUntilChanged()
+            .debounce(1.seconds)
+            .collectIn(viewModelScope) { (title, content) ->
+                notesRepository.update(noteId, title, content)
+            }
     }
 
     fun onTitleChange(title: String) {
@@ -62,21 +69,43 @@ internal class NoteScreenViewModel(
     }
 
     fun onDeleteClick() {
-        viewModelScope.launch {
-            notesRepository.delete(noteId)
-            navigator.popBackStack()
-        }
+        navigator.showOverlay(
+            overlay = BottomSheet.Confirm(
+                text = { stringResource(Res.string.note_delete_confirmation) },
+                onConfirm = {
+                    viewModelScope.launch {
+                        notesRepository.delete(noteId)
+                        navigator.popBackStack()
+                    }
+                }
+            ),
+        )
     }
 
     fun onBackClick() {
-        val current = _uiState.value as? NoteUiState.Editing
-        if (current != null) {
-            viewModelScope.launch {
-                notesRepository.update(noteId, current.title, current.content)
-                navigator.popBackStack()
+        navigator.popBackStack()
+    }
+
+    private suspend fun updateToActual(
+        state: NoteUiState.Editing,
+    ) {
+        notesRepository.update(
+            id = noteId,
+            title = state.title,
+            content = state.content,
+        )
+    }
+
+    @OptIn(DelicateCoroutinesApi::class)
+    override fun onCleared() {
+        when (val state = uiState.value) {
+            is NoteUiState.Editing -> GlobalScope.launch {
+                updateToActual(state)
             }
-        } else {
-            navigator.popBackStack()
+
+            is NoteUiState.Loading -> Unit
         }
+
+        super.onCleared()
     }
 }
